[api/libwgbuilder] Make DNS server optional
**Why is the change necessary?**

In some cases clients need more than one wireguard tunnel. If this is
the case, the last configured wireguard interfaces defines the tunnel
for the DNS requests. This might be not the desired behavior.

**How does it address the issue?**
Make the DNS server optional for the client so that one can choose over
which tunnel the DNS requests should be sent.