Fix login not possible without random token content
The ApiKey extraction extracts the api key in each request from the
headers. It returns a HTTP bad request code if the Authorization header
is not set. This extraction is also made for the login mutation. A
validation of this token is only done in the UserGuard and the AdminGuard.
This is why the token content for the login mutation can be any random
value. It is just validated if there is a token but not validating the
token at this point.

Solution: Just check if the Authorization header is set and extract the
token from it. If the header is not set an empty api key is set. This
will fail while validating the token later.

Closes #209