Remove redundant ServerGuard check
The ServerGuard is already checked on the server query.
Because of this it is not necessary to check it again on the
configuration field. This additional check is currently implemented and
makes it impossible to get the configuration of a server as an
administrator.

Solution: Remove the duplicate check on the configuration field.